+++ CHANGELOG.md
@@ -2,6 +2,53 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="2.4.2"></a>
+## [2.4.2](https://github.com/ng-select/ng-select/compare/v2.4.1...v2.4.2) (2018-08-12)
+
+
+### Bug Fixes
+
+* listen to `touchstart` events for outside click. ([70fcd66](https://github.com/ng-select/ng-select/commit/70fcd66))
+* remove circular dependency between dropdown and select cmp ([#741](https://github.com/ng-select/ng-select/issues/741)) ([7520e85](https://github.com/ng-select/ng-select/commit/7520e85)), closes [#668](https://github.com/ng-select/ng-select/issues/668)
+
+
+
+<a name="2.4.1"></a>
+## [2.4.1](https://github.com/ng-select/ng-select/compare/v2.4.0...v2.4.1) (2018-08-09)
+
+
+### Bug Fixes
+
+* allow to reposition input cursor using mouse ([25de23f](https://github.com/ng-select/ng-select/commit/25de23f)), closes [#644](https://github.com/ng-select/ng-select/issues/644)
+
+
+
+<a name="2.4.0"></a>
+# [2.4.0](https://github.com/ng-select/ng-select/compare/v2.3.6...v2.4.0) (2018-08-08)
+
+
+### Bug Fixes
+
+* **tagging:** don't show tag if it exists among items or selected ([d5e97af](https://github.com/ng-select/ng-select/commit/d5e97af)), closes [#693](https://github.com/ng-select/ng-select/issues/693)
+* accept spaces while tagging without dropdown ([4320098](https://github.com/ng-select/ng-select/commit/4320098)), closes [#676](https://github.com/ng-select/ng-select/issues/676)
+
+
+### Features
+
+* support tabindex ([75e9ac2](https://github.com/ng-select/ng-select/commit/75e9ac2)), closes [#680](https://github.com/ng-select/ng-select/issues/680)
+
+
+
+<a name="2.3.6"></a>
+## [2.3.6](https://github.com/ng-select/ng-select/compare/v2.3.5...v2.3.6) (2018-07-30)
+
+
+### Bug Fixes
+
+* focus select after unselecting single item ([1bc3f02](https://github.com/ng-select/ng-select/commit/1bc3f02)), closes [#661](https://github.com/ng-select/ng-select/issues/661)
+
+
+
 <a name="2.3.5"></a>
 ## [2.3.5](https://github.com/ng-select/ng-select/compare/v2.3.4...v2.3.5) (2018-07-18)